chk-followups/hash_by_ref: Use prev results as cache
In case of previous results use, much time is spent in the comparison of
the tracking ranges for the previous results and current request, due to
'hash_by_ref()', because it is intensively called but the function uses
'git' internally.  This commit optimizes the call by utilizing another
previous results cache.  This change reduced repeated 5.4.y tracking
time (no change in the tracking ranges between the previous and current)
from 4m23s to 2.971s.

Signed-off-by: SeongJae Park <sjpark@amazon.de>