chk-followups: Let the first commit in rule not the followee
The upstream fixes is not always merged in the root cause commit but can
be merged in different name.  Current ignore rule cannot applied to the
case.  Therefore, this commit extend the rule so that the first commit
of each rule is not only the followee, but the ignorance trigger commit.

Signed-off-by: SeongJae Park <sjpark@amazon.de>